+++ StarkBank/StarkBank/UtilityPayment/Log.cs
@@ -20,7 +20,7 @@ namespace StarkBank
         ///     <item>Payment [UtilityPayment]: UtilityPayment entity to which the log refers to.</item>
         ///     <item>Errors [list of strings]: list of errors linked to this UtilityPayment event.</item>
         ///     <item>Type [string]: type of the UtilityPayment event which triggered the log creation. ex: "processing" or "success"</item>
-        ///     <item>Created [DateTime]: creation datetime for the log. ex: DateTime.new(2020, 3, 10, 10, 30, 0, 0)</item>
+        ///     <item>Created [DateTime]: creation datetime for the log. ex: new DateTime(2020, 3, 10, 10, 30, 0, 0)</item>
         /// </list>
         /// </summary>
         public class Log : Utils.Resource
@@ -43,7 +43,7 @@ namespace StarkBank
             ///     <item>payment [UtilityPayment]: UtilityPayment entity to which the log refers to.</item>
             ///     <item>errors [list of strings]: list of errors linked to this UtilityPayment event.</item>
             ///     <item>type [string]: type of the UtilityPayment event which triggered the log creation. ex: "processing" or "success"</item>
-            ///     <item>created [DateTime]: creation datetime for the log. ex: DateTime.new(2020, 3, 10, 10, 30, 0, 0)</item>
+            ///     <item>created [DateTime]: creation datetime for the log. ex: new DateTime(2020, 3, 10, 10, 30, 0, 0)</item>
             /// </list>
             /// </summary>
             public Log(string id, DateTime created, string type, List<string> errors, UtilityPayment payment) : base(id)