+++ StarkBank/StarkBank/Utils/Resource.cs
@@ -24,24 +24,32 @@ namespace StarkBank.Utils
             string spacer = "  ";
             string newLine = Environment.NewLine;
             List<string> fields = new List<string>();
-            foreach (KeyValuePair<string, object> entry in dict)
-            {
+            foreach (KeyValuePair<string, object> entry in dict) {
                 string stringValue;
-                if (entry.Value is null)
-                {
+                if (entry.Value is null) {
                     stringValue = "null";
                 }
-                else if (entry.Value is List<string> stringList)
-                {
+                else if (entry.Value is List<string> stringList) {
                     stringValue = "{ " + string.Join(", ", stringList) + " }";
                 }
-                else if (entry.Value is List<Dictionary<string, object>> dictList)
+                else if (entry.Value is List<Dictionary<string, object>> objectDictList) {
+                    List<string> stringDictList = new List<string>();
+                    foreach (Dictionary<string, object> subDict in objectDictList) {
+                        List<string> stringDictSubList = new List<string>();
+                        foreach (KeyValuePair<string, object> subEntry in subDict) {
+                            stringDictSubList.Add("{ " + subEntry.Key + ", " + subEntry.Value.ToString() + " }");
+                        }
+                        stringDictList.Add("{ " + string.Join(", ", stringDictSubList) + " }");
+                    }
+                    stringValue = "{ " + string.Join(", ", stringDictList) + " }";
+                }
+                else if (entry.Value is List<Dictionary<string, string>> dictList)
                 {
                     List<string> stringDictList = new List<string>();
-                    foreach (Dictionary<string, object> subDict in dictList)
+                    foreach (Dictionary<string, string> subDict in dictList)
                     {
                         List<string> stringDictSubList = new List<string>();
-                        foreach (KeyValuePair<string, object> subEntry in subDict)
+                        foreach (KeyValuePair<string, string> subEntry in subDict)
                         {
                             stringDictSubList.Add("{ " + subEntry.Key + ", " + subEntry.Value.ToString() + " }");
                         }
@@ -49,8 +57,7 @@ namespace StarkBank.Utils
                     }
                     stringValue = "{ " + string.Join(", ", stringDictList) + " }";
                 }
-                else
-                {
+                else {
                     stringValue = entry.Value.ToString();
                 }
                 string[] splitValue = stringValue.Split(