+++ StarkBank/StarkBankTests/TransferLog.cs
@@ -24,13 +24,13 @@ namespace StarkBankTests
             Assert.True(logs.First().ID != logs.Last().ID);
             foreach (Transfer.Log log in logs)
             {
-                Console.WriteLine(log);
+                TestUtils.Log(log);
                 Assert.NotNull(log.ID);
                 Assert.Equal("failed", log.Type);
             }
             Transfer.Log getLog = Transfer.Log.Get(id: logs.First().ID);
             Assert.Equal(getLog.ID, logs.First().ID);
-            Console.WriteLine(getLog);
+            TestUtils.Log(getLog);
         }
     }
 }