+++ StarkBank/StarkBankTests/UtilityPaymentLog.cs
@@ -22,13 +22,13 @@ namespace StarkBankTests
             Assert.True(logs.First().ID != logs.Last().ID);
             foreach (UtilityPayment.Log log in logs)
             {
-                Console.WriteLine(log);
+                TestUtils.Log(log);
                 Assert.NotNull(log.ID);
                 Assert.Equal("success", log.Type);
             }
             UtilityPayment.Log getLog = UtilityPayment.Log.Get(id: logs.First().ID);
             Assert.Equal(getLog.ID, logs.First().ID);
-            Console.WriteLine(getLog);
+            TestUtils.Log(getLog);
         }
     }
 }