+++ StarkBank/StarkBankTests/Utils.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+
+namespace StarkBankTests
+{
+    public static class TestUtils
+    {
+        internal static void Log(object log)
+        {
+            //Console.WriteLine(log);
+        }
+    }
+}