+++ StarkBank/StarkBankTests/Webhook.cs
@@ -21,7 +21,7 @@ namespace StarkBankTests
             Assert.Equal(getWebhook.ID, webhook.ID);
             Webhook deleteWebhook = Webhook.Delete(id: webhook.ID);
             Assert.Equal(deleteWebhook.ID, webhook.ID);
-            Console.WriteLine(deleteWebhook);
+            TestUtils.Log(deleteWebhook);
         }
 
         [Fact]
@@ -30,7 +30,7 @@ namespace StarkBankTests
             List<Webhook> webhooks = Webhook.Query().ToList();
             foreach (Webhook webhook in webhooks)
             {
-                Console.WriteLine(webhook);
+                TestUtils.Log(webhook);
                 Assert.NotNull(webhook.ID);
                 Assert.NotNull(webhook.Url);
                 Assert.NotNull(webhook.Subscriptions);
@@ -38,7 +38,7 @@ namespace StarkBankTests
             }
         }
 
-        private Webhook Example()
+        internal static Webhook Example()
         {
             return new Webhook(
                 url: "https://webhook.site/60e9c18e-4b5c-4369-bda1-ab5fcd8e1b29",